// types/starbucksTypes.tsx
export interface headerMenu {
    id : number;
    name : string;
    link : string;
}

export interface loginData {
    email: string;
    password: string;
}

export interface productDataType {
    id: number;
    name: string;
    price: number;
    description : string;
    thumbnail: string;
    size: string;
    season: string;
}